Adapt flux deployment and give it a service account etc.
This commit adapts the generated flux manifests so that they fit the
layered structure here. Flux is set up with a restricted service
account (*not very restricted) and pointed at the specific bit of git
repo it should sync for the app in question.

It also moves the cuttlefacts app into its own directory and
namespace, rather than putting directly in app/, to imply multitenancy
on this layer.