Sync policy layer from infrastructural Flux
The general principle of the layering is that each layer instantiates
the one above it. E.g., substructure syncs the definitions for
platform.

The difference between platform and policy is mechanism vs
particulars, but it's a thin distinction. To keep things moving, I'm
going to conflate them with respect to instantiation, and just tell
the Flux in infrastructure to sync policy/ as well.