Provide ingress for our app
This commit adds an ingress controller to the platform layer, and an
ingress (routing rule) for the cuttlefacts app.

It's arguable whether the latter should go into the policy, since the
routing rules might be considered part of the app. For our purposes
here, I'm considering it part of the domain of whoever is in charge of
provisioning apps.